+++ config/api/category/parameters.php
@@ -4,19 +4,19 @@ declare(strict_types=1);
 
 return [
     'collection' => [
-        'include_sub_categories' => [
-            'field' => 'include_sub_categories',
-            'title' => 'category/parameters.title-include_sub_categories',
-            'description' => 'category/parameters.description-include_sub_categories',
+        'include-subcategories' => [
+            'field' => 'include-subcategories',
+            'title' => 'category/parameters.title-include-subcategories',
+            'description' => 'category/parameters.description-include-subcategories',
             'type' => 'boolean',
             'required' => false
         ]
     ],
     'item' => [
-        'include_sub_categories' => [
-            'field' => 'include_sub_categories',
-            'title' => 'category/parameters.title-include_sub_categories',
-            'description' => 'category/parameters.description-include_sub_categories',
+        'include-subcategories' => [
+            'field' => 'include-subcategories',
+            'title' => 'category/parameters.title-include-subcategories',
+            'description' => 'category/parameters.description-include-subcategories',
             'type' => 'boolean',
             'required' => false
         ]