+++ resources/lang/en/category/parameters.php
@@ -3,6 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'title-include_sub_categories' => 'Include subcategories?',
-    'description-include_sub_categories' => 'Optionally include the subcategories for each category'
+    'title-include-subcategories' => 'Include subcategories?',
+    'description-include-subcategories' => 'Optionally include the subcategories for each category'
 ];