+++ resources/lang/en/resource-type/parameters.php
@@ -3,6 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'title-include_resources' => 'Include resources?',
-    'description-include_resources' => 'Optionally include the resources for each resource type'
+    'title-include-resources' => 'Include resources?',
+    'description-include-resources' => 'Optionally include the resources for each resource type'
 ];