+++ resources/lang/en/sortable.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'title' => 'Sorting options',
+    'description' => 'Sortable fields are all sortable in ascending and descending order, example sort=field1:asc|field2:desc'
+];