+++ routes/api/public-summary-routes.php
@@ -14,6 +14,26 @@ Route::group(
     ],
     function () {
         Route::get(
+            'summary/resource-types',
+            'SummaryResourceTypeController@index'
+        );
+
+        Route::options(
+            'summary/resource-types',
+            'SummaryResourceTypeController@optionsIndex'
+        );
+
+        Route::get(
+            'summary/resource-types/{resource_type_id}/resources',
+            'SummaryResourceController@index'
+        );
+
+        Route::options(
+            'summary/resource-types/{resource_type_id}/resources',
+            'SummaryResourceController@optionsIndex'
+        );
+
+        Route::get(
             'summary/resource-types/{resource_type_id}/items',
             'SummaryResourceTypeItemController@index'
         );
@@ -22,7 +42,6 @@ Route::group(
             'summary/resource-types/{resource_type_id}/items',
             'SummaryResourceTypeItemController@optionsIndex'
         );
-
 
         Route::get(
             'summary/resource-types/{resource_type_id}/resources/{resource_id}/items',